Updated module augmentation for undo package.
