+++ packages/ckeditor5-undo/src/index.ts
@@ -11,3 +11,5 @@ export { default as Undo } from './undo';
 export { default as UndoEditing } from './undoediting';
 export { default as UndoUi } from './undoui';
 export type { default as UndoCommand } from './undocommand';
+
+import './augmentation';
